requirements.txt
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Sitemap==0.3.0
Flask-SocketIO==3.3.1
gcloud==0.17.0
googleapis-common-protos==1.5.8
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jws==0.1.3
lxml==4.3.1
MarkupSafe==1.1.0
oauth2client==3.0.0
Pillow==5.4.1
protobuf==3.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.4.3
Pyrebase==3.0.27
python-engineio==3.4.2
python-jwt==2.0.1
python-socketio==3.1.2
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
stripe==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
